Run `forge fmt`

+++ src/fundraising/ShowtimeBakeSale.sol
@@ -2,19 +2,16 @@
 
 pragma solidity ^0.8.7;
 
-import { Ownable } from "@openzeppelin/contracts/access/Ownable.sol";
-import { PaymentSplitter } from "@openzeppelin/contracts/finance/PaymentSplitter.sol";
-import { IERC1155 } from "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
+import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
+import {PaymentSplitter} from "@openzeppelin/contracts/finance/PaymentSplitter.sol";
+import {IERC1155} from "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
 
-import { ShowtimeMTReceiver } from "./ShowtimeMTReceiver.sol";
+import {ShowtimeMTReceiver} from "./ShowtimeMTReceiver.sol";
 
 interface IShowtimeV1Market {
-    function createSale(
-        uint256 _tokenId,
-        uint256 _quantity,
-        uint256 _price,
-        address _currency
-    ) external returns (uint256 listingId);
+    function createSale(uint256 _tokenId, uint256 _quantity, uint256 _price, address _currency)
+        external
+        returns (uint256 listingId);
 
     function cancelSale(uint256 _listingId) external;
 }
@@ -33,23 +30,20 @@ interface IShowtimeV1Market {
 contract ShowtimeBakeSale is PaymentSplitter, ShowtimeMTReceiver, Ownable {
     IShowtimeV1Market public immutable showtimeMarket;
 
-    constructor(
-        address _showtimeMT,
-        address _showtimeMarket,
-        address[] memory payees,
-        uint256[] memory shares_
-    ) PaymentSplitter(payees, shares_) ShowtimeMTReceiver(_showtimeMT) {
+    constructor(address _showtimeMT, address _showtimeMarket, address[] memory payees, uint256[] memory shares_)
+        PaymentSplitter(payees, shares_)
+        ShowtimeMTReceiver(_showtimeMT)
+    {
         showtimeMarket = IShowtimeV1Market(_showtimeMarket);
 
         IERC1155(_showtimeMT).setApprovalForAll(_showtimeMarket, true);
     }
 
-    function createSale(
-        uint256 _tokenId,
-        uint256 _quantity,
-        uint256 _price,
-        address _currency
-    ) external onlyOwner returns (uint256 listingId) {
+    function createSale(uint256 _tokenId, uint256 _quantity, uint256 _price, address _currency)
+        external
+        onlyOwner
+        returns (uint256 listingId)
+    {
         listingId = showtimeMarket.createSale(_tokenId, _quantity, _price, _currency);
     }
 
@@ -58,7 +52,7 @@ contract ShowtimeBakeSale is PaymentSplitter, ShowtimeMTReceiver, Ownable {
     }
 
     function withdraw(uint256 _tokenId, address to) external onlyOwner {
-        uint howMany = IERC1155(showtimeMT).balanceOf(address(this), _tokenId);
+        uint256 howMany = IERC1155(showtimeMT).balanceOf(address(this), _tokenId);
         IERC1155(showtimeMT).safeTransferFrom(address(this), to, _tokenId, howMany, "");
     }
 }
